chore(cleanup): refactor for cleanup
