+++ src/graphql/Auth.ts
@@ -1,10 +1,6 @@
-import * as bcrypt from "bcryptjs";
-import * as jwt from "jsonwebtoken";
-
 import { stringArg, nonNull, extendType } from "nexus";
 
-import { APP_SECRET_KEY } from "../utils/auth";
-import { userLoginSchema, userSignUpSchema } from "../utils/validators";
+import { loginResolver, signupResolver } from "./resolvers";
 
 export const AuthMutation = extendType({
     type: "Mutation",
@@ -15,26 +11,7 @@ export const AuthMutation = extendType({
                 email: nonNull(stringArg()),
                 password: nonNull(stringArg()),
             },
-            async resolve(parent, args, context) {
-                const user = await context.prisma.user.findUnique({
-                    where: { email: args.email },
-                });
-                if (!user) {
-                    throw new Error("Invalid credentials");
-                }
-                const valid = await bcrypt.compare(
-                    args.password,
-                    user.password,
-                );
-                if (!valid) {
-                    throw new Error("Invalid credentials");
-                }
-                const token = jwt.sign({ userId: user.id }, APP_SECRET_KEY);
-                return {
-                    token,
-                    user,
-                };
-            },
+            resolve: loginResolver
         });
         t.nonNull.field("signup", {
             type: "AuthPayload",
@@ -43,23 +20,8 @@ export const AuthMutation = extendType({
                 password: nonNull(stringArg()),
                 name: nonNull(stringArg()),
             },
-            async resolve(parent, args, context) {
-                // TODO: Add inout validation e.g. for email and password use Yup
-                const { email, name } = args;
-                const password = await bcrypt.hash(args.password, 10);
-                const user = await context.prisma.user.create({
-                    data: {
-                        email,
-                        name,
-                        password
-                    }
-                })
-                const token = jwt.sign({ userId: user.id }, APP_SECRET_KEY);
-                return {
-                    token,
-                    user
-                }
-            }
+            resolve: signupResolver
+           
         });
     },
 });
