+++ src/graphql/resolvers/index.ts
@@ -1,1 +1,2 @@
 export * from "./track";
+export * from "./auth";
